Missing SCs, units with bad destination data
There were a number of tiles that didn't have SCs, resulting in some of
the provinces appearing to have more units than SCs. This was an issue
with the map data and is now fixed.

When setting the unit destination, we were setting it when creating the
animation, meaning that when we went to create the next animation we
would find the unit we already made an animation for.